refactor(app): create coordinating component for offset banner priority
